+++ packages/@storefront/breadcrumbs/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [patch]
+### Fixed
+- SF-1248: Added a `shouldUpdate` function to the `refinement-crumbs` component.
+  - The addition of this function allows for the component to update as needed and have the latest data from the application store.
+
 ## [2.4.0] - 2019-01-30
 ### Changed
 - Update `@storefront/core` to 2.4.0.